Adding .zsh configuration file
This expects a handful of tools to be installed in order to work out of
the box.  I personally use nvim, fzf, exa, fd, and bat/batcat.  I also
use Rust and python, so it expects those to be installed.  I comment out
all of the oh-my-zsh requirements, and labelled them with the
requirement.